relctl: one hermetic git fixture, replacing three that were not
The Go port grew three copies of the same builder and none of them
neutralised the ambient git configuration. hack/release/next_version_test.go
solved this for the shell suite and said why; the port dropped the guard
and reintroduced exactly the problem it describes.

Reproduced before fixing. With commit.gpgsign=true set globally:

  fatal: failed to write commit object
  FAIL

That is 76 resolver cases, the classify suite and hundreds of
differential comparisons going red at once, on a maintainer's machine
only, for a reason nothing in the output points at.

One builder now, with GIT_CONFIG_GLOBAL and GIT_CONFIG_SYSTEM pointed at
/dev/null and the identity supplied through the environment rather than
`git config`, so a fixture needs no configuration at all and cannot
inherit a template or hooks path before its first commit. Verified
against a config setting commit.gpgsign, tag.gpgSign, core.hooksPath and
init.templateDir together: both packages pass.
